app/pyproject.toml
[project]
name = "amanzon"
version = "0.1.0"
description = "A modern Django e-commerce application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.9",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "razorpay>=2.0.0",
]
